visual-product-search/requirements.txt
opencv-python>=4.8
numpy>=1.24
faiss-cpu>=1.7
pytest>=7.0
